requirements.txt
prometheus-client==0.6.0
python-jenkins==1.7.0
argparse==1.4.0
requests==2.22.0
xmltodict==0.12.0
